Module parry2d_f64::math

Compilation flags dependent aliases for mathematical types.

Structs

SimdBool

An SIMD boolean structure associated to wide::f64x4 that implements all the relevant traits from simba.

SimdReal

A wrapper type of wide::f64x4 that implements all the relevant traits from num and simba.

Constants

DEFAULT_EPSILON

The default tolerance used for geometric operations.

DIM

The dimension of the space.

SIMD_LAST_INDEX

SIMD_WIDTH - 1

SIMD_WIDTH

The number of lanes of a SIMD number.

TWO_DIM

The dimension of the space multiplied by two.

Type Definitions

AngDim

The dimension of the rotations.

AngVector

The angular vector type.

AngularInertia

The angular inertia of a rigid body.

CrossMatrix

A matrix that represent the cross product with a given vector.

Dim

The dimension of the ambient space.

Isometry

The transformation matrix type.

Matrix

The matrix type.

Orientation

The orientation type.

Point

The point type.

PrincipalAngularInertia

The principal angular inertia of a rigid body.

Real

The scalar type used throughout this crate.

Rotation

The rotation matrix type.

SdpMatrix

A 2D symmetric-definite-positive matrix.

SpacialVector

A vector with a dimension equal to the maximum number of degrees of freedom of a rigid body.

Translation

The translation type.

UnitVector

The unit vector type.

Vector

The vector type.