Type Alias parry2d_f64::math::SdpMatrix

pub type SdpMatrix<N> = SdpMatrix2<N>;

A 2D symmetric-definite-positive matrix.

Aliased Type

struct SdpMatrix<N> {
    pub m11: N,
    pub m12: N,
    pub m22: N,
}

Fields
m11: N

The component at the first row and first column of this matrix.

m12: N

The component at the first row and second column of this matrix.

m22: N

The component at the second row and second column of this matrix.

Implementations

impl<N: SimdRealField + Copy> SdpMatrix2<N>

pub fn new(m11: N, m12: N, m22: N) -> Self

A new SDP 2x2 matrix with the given components.

Because the matrix is symmetric, only the lower off-diagonal component is required.

pub fn from_sdp_matrix(mat: Matrix2<N>) -> Self

Build an SdpMatrix2 structure from a plain matrix, assuming it is SDP.

No check is performed to ensure mat is actually SDP.

pub fn zero() -> Self

Create a new SDP matrix filled with zeros.

pub fn diagonal(val: N) -> Self

Create a new SDP matrix with its diagonal filled with val, and its off-diagonal elements set to zero.

pub fn add_diagonal(&mut self, elt: N) -> Self

Adds val to the diagonal components of self.

pub fn inverse_unchecked(&self) -> Self

Compute the inverse of this SDP matrix without performing any inversibility check.

pub fn into_matrix(self) -> Matrix2<N>

Convert this SDP matrix to a regular matrix representation.