Type Alias parry2d_f64::shape::HeightField

pub type HeightField = GenericHeightField<DefaultStorage>;

A 2D heightfield.

Aliased Type

struct HeightField { /* private fields */ }

Implementations

impl HeightField

pub fn new(heights: DVector<Real>, scale: Vector<Real>) -> Self

Creates a new 2D heightfield with the given heights and scale factor.

impl HeightField

pub fn to_polyline(&self) -> (Vec<Point2<Real>>, Vec<[u32; 2]>)

Rasterize this heightfield as a (potentially discontinuous) polyline.

impl<Storage: HeightFieldStorage> GenericHeightField<Storage>

pub fn aabb(&self, pos: &Isometry<Real>) -> Aabb

Computes the world-space Aabb of this heightfield, transformed by pos.

pub fn local_aabb(&self) -> Aabb

Computes the local-space Aabb of this heightfield.

impl<Storage: HeightFieldStorage> GenericHeightField<Storage>

pub fn bounding_sphere(&self, pos: &Isometry<Real>) -> BoundingSphere

Computes the world-space bounding sphere of this height-field, transformed by pos.

pub fn local_bounding_sphere(&self) -> BoundingSphere

Computes the local-space bounding sphere of this height-field.

impl<Storage: HeightFieldStorage> GenericHeightField<Storage>

pub fn num_cells(&self) -> usize

The number of cells of this heightfield.

pub fn heights(&self) -> &Storage::Heights

The height at each cell endpoint.

pub fn scale(&self) -> &Vector<Real>

The scale factor applied to this heightfield.

pub fn set_scale(&mut self, new_scale: Vector<Real>)

Sets the scale factor applied to this heightfield.

pub fn scaled(self, scale: &Vector<Real>) -> Self

Returns a scaled version of this heightfield.

pub fn root_aabb(&self) -> &Aabb

The Aabb of this heightfield.

pub fn cell_width(&self) -> Real

The width of a single cell of this heightfield.

pub fn unit_cell_width(&self) -> Real

The width of a single cell of this heightfield, without taking the scale factor into account.

pub fn start_x(&self) -> Real

The left-most x-coordinate of this heightfield.

pub fn cell_at_point(&self, pt: &Point2<Real>) -> Option<usize>

Index of the cell a point is on after vertical projection.

pub fn height_at_point(&self, pt: &Point2<Real>) -> Option<Real>

Height of the heightfield a the given point after vertical projection on the heightfield surface.

pub fn segments<'a>(&'a self) -> impl Iterator<Item = Segment> + 'a

Iterator through all the segments of this heightfield.

pub fn segment_at(&self, i: usize) -> Option<Segment>

The i-th segment of the heightfield if it has not been removed.

pub fn set_segment_removed(&mut self, i: usize, removed: bool)

Mark the i-th segment of this heightfield as removed or not.

pub fn is_segment_removed(&self, i: usize) -> bool

Checks if the i-th segment has been removed.

pub fn unclamped_elements_range_in_local_aabb( &self, aabb: &Aabb ) -> Range<isize>

The range of segment ids that may intersect the given local Aabb.

pub fn map_elements_in_local_aabb( &self, aabb: &Aabb, f: &mut impl FnMut(u32, &Segment) )

Applies f to each segment of this heightfield that intersects the given aabb.