Type Alias parry2d_f64::math::SdpMatrix

pub type SdpMatrix<N> = SdpMatrix2<N>;

A 2D symmetric-definite-positive matrix.

Aliased Type

struct SdpMatrix<N> {
    pub m11: N,
    pub m12: N,
    pub m22: N,
}

Fields
m11: N

The component at the first row and first column of this matrix.

m12: N

The component at the first row and second column of this matrix.

m22: N

The component at the second row and second column of this matrix.